refactor(metrics): migrate remaining modules to @metric registry (#472)
Completes the spec-to-registry migration started in db5d8b9: the
remaining family modules drop their legacy __metric_specs__ tuples for
the class-based @metric decorator, leaving the registry as the single
spec source.

With no legacy tuples left, _all_specs() folds back onto
_load_module_specs (one requires-validation path, set-based dedup)
instead of a parallel hand-rolled walk. Series and scalar diagnostics
also gain their true input_shape, which the tuple form defaulted to
PANEL.

Refs #472
