+++ factrix/metrics/_primitives/__init__.py
@@ -1,15 +1,19 @@
 from factrix.metrics._primitives.compute_caar import compute_caar
 from factrix.metrics._primitives.compute_event_returns import compute_event_returns
 from factrix.metrics._primitives.compute_fm_betas import compute_fm_betas
+from factrix.metrics._primitives.compute_group_returns import compute_group_returns
 from factrix.metrics._primitives.compute_ic import compute_ic
 from factrix.metrics._primitives.compute_mfe_mae import compute_mfe_mae
+from factrix.metrics._primitives.compute_spread_series import compute_spread_series
 from factrix.metrics._primitives.compute_ts_betas import compute_ts_betas
 
 __all__ = [
     "compute_caar",
     "compute_event_returns",
     "compute_fm_betas",
+    "compute_group_returns",
     "compute_ic",
     "compute_mfe_mae",
+    "compute_spread_series",
     "compute_ts_betas",
 ]
