+++ factrix/metrics/trend.py
@@ -22,29 +22,28 @@ from factrix._axis import (
     Aggregation,
     DataStructure,
     FactorDensity,
+    InputShape,
     SEMethod,
     TestMethod,
 )
-from factrix._metric_index import MetricSpec, cell
+from factrix._metric_index import cell
 from factrix._results import MetricResult
 from factrix._stats import _adf, _p_value_from_t
+from factrix.metrics import metric
 from factrix.metrics._helpers import _short_circuit_output
-
-__metric_specs__ = (
-    MetricSpec(
-        name="ic_trend",
-        cell=cell(None, FactorDensity.DENSE, structure=DataStructure.TIMESERIES),
-        aggregation=Aggregation.TS_ONLY,
-        test_method=TestMethod.RANK,
-        se_method=SEMethod.BUILT_IN,
-    ),
-)
 
 __all__ = [
     "ic_trend",
 ]
 
 
+@metric(
+    cell=cell(None, FactorDensity.DENSE, structure=DataStructure.TIMESERIES),
+    aggregation=Aggregation.TS_ONLY,
+    test_method=TestMethod.RANK,
+    se_method=SEMethod.BUILT_IN,
+    input_shape=InputShape.SERIES,
+)
 def ic_trend(
     series: pl.DataFrame,
     value_col: str = "value",
